+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @abstract-money/cli
 
+## 0.3.4
+
+### Patch Changes
+
+- [`21f8b18`](https://github.com/AbstractSDK/abstract.js/commit/21f8b18a813b44ff0399c422bbec971a328f34a2) Thanks [@dalechyn](https://github.com/dalechyn)! - Bumped @cosmjs dependencies versions.
+
+## 0.3.3
+
+### Patch Changes
+
+- [#96](https://github.com/AbstractSDK/abstract.js/pull/96) [`fc1d460`](https://github.com/AbstractSDK/abstract.js/commit/fc1d46077f5f242c02a6d671920b1906e3dbd5de) Thanks [@dalechyn](https://github.com/dalechyn)! - Return full query result from generated queries.
+
 ## 0.3.2
 
 ### Patch Changes